+++ relocate.js
@@ -24,23 +24,17 @@ fs.copyFile("./dist/core/actor.d.ts", "./actor.d.ts", error => {
     }
 });
 
-fs.copyFile("./dist/core/broadcaster.js", "./broadcaster.js", error => {
+fs.copyFile("./dist/web_modules/broadcaster.js", "./broadcaster.js", error => {
     if (error) {
         console.log(error);
     }
 });
-fs.copyFile("./dist/core/broadcaster.d.ts", "./broadcaster.d.ts", error => {
+fs.copyFile("./dist/web_modules/broadcaster.d.ts", "./broadcaster.d.ts", error => {
     if (error) {
         console.log(error);
     }
 });
-
-fs.copyFile("./dist/core/fetch.js", "./fetch.js", error => {
-    if (error) {
-        console.log(error);
-    }
-});
-fs.copyFile("./dist/core/fetch.d.ts", "./fetch.d.ts", error => {
+fs.copyFile("./node_modules/wwibs/broadcaster-worker.min.js", "./broadcaster-worker.min.js", error => {
     if (error) {
         console.log(error);
     }